Add token object in addition to accessToken, to expose idToken and more
Before the change, the hook exposed acessToken only. But by using
auth0-spa-js, we're getting also the idToken, which is needed for
cases we use Auth0 as an authorisation service that defines the rules
for accessing data in our server / database.

This change allows externalizing other parameters in Auth0's response,
such as idToken and the expiration of the token.